ai_service/requirements.txt
gunicorn
flask==3.0.3
python-dotenv==1.0.1
google-generativeai==0.8.3
flask-cors==5.0.0
werkzeug==3.0.1